requirements.txt
attrs==23.1.0
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
fiona==1.9.5
fonttools==4.43.1
geographiclib==2.0
geopandas==0.14.0
geopy==2.4.0
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pip==23.1.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.6
seaborn==0.13.0
setuptools==65.5.0
shapely==2.0.2
six==1.16.0
tzdata==2023.3
